+++ TeamTalk-Swift/TeamTalk-Swift/Exist/Voice/manager/RawAudioDataPlayer.m
@@ -130,7 +130,7 @@ static void AudioPlayerAQInputCallback(void *input, AudioQueueRef inQ, AudioQueu
                 [self removeEmptyBuffer:outQB];
                 
                 NSUInteger readLength = (lengthLeft > EVERY_READ_LENGTH) ? EVERY_READ_LENGTH : lengthLeft;
-                outQB->mAudioDataByteSize = readLength;
+                outQB->mAudioDataByteSize = (UInt32)readLength;
                 Byte *audioData = (Byte *)outQB->mAudioData;
                 memcpy(audioData, [mPcmData bytes] + readedBytes, readLength);
                 readedBytes += readLength;
